+++ CHANGELOG.md
@@ -20,6 +20,7 @@ For questions and support requests, please use
 * Hashtags are now handled independent of the social media plugins
 * Show the logo of the social media network in the social media post
 * Keep existing posts to avoid empty screens on errors
+* Add configuration option to adjust event times (not just the dates)
 
 ### Fixed Bugs